+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.7.0] - 2019-07-26
+### Added
+- Support to handle Fennec nightlies from ESR branch
+- Support to handle `fennec-nightly-latest` locations in Fennec nightlies
+
 ## [3.6.2] - 2019-07-09
 ### Fixed
 - Fennec beta new pattern (68.XbN)